Add ability to change color of card frame
Add the ability to change the color of the frame that surrounds the phone while scanning a card. This can be done by overriding the resource value `wocr_accent` in the app.

Test plan: Add this line to demo/res/values/colors.xml


    <color name="wocr_accent">@color/colorAccent</color>

Expected behavior: The lines around the card have changed from green to blue